requirements.txt
langchain==0.0.239
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.25.0
qdrant-client==1.3.1
openai==0.27.8
tiktoken
